feat: support ash (alpine)
rewrote the scripts to use features that are available in both ash and
bash. I think the largest change is the switch to no longer use bash
arrays, and the use of [ over [[, although it seems alpine-latest
has some [[ executable although i'm not sure if that would exist on
older alpine versions we'd want to support.